+++ newtutils/__init__.py
@@ -44,7 +44,8 @@ from .files import (
     save_json_to_file,
     read_csv_from_file,
     save_csv_to_file,
-    setup_logging, cleanup_logging,
+    setup_logging,
+    cleanup_logging,
 )
 
 # SQL
@@ -84,9 +85,10 @@ __all__ = [
     "convert_str_to_json",
     "read_json_from_file",
     "save_json_to_file",
-    # "read_csv_from_file",
-    # "save_csv_to_file",
-    "setup_logging", "cleanup_logging",
+    "read_csv_from_file",
+    "save_csv_to_file",
+    "setup_logging",
+    "cleanup_logging",
     # SQL ----------
     "db_delayed_close",
     "sql_execute_query",